+++ backend/internal/storage/postgres/feedback_comments.go
@@ -10,7 +10,7 @@ import (
 )
 
 // gets all feedback comments on a student work
-func (db *DB) GetFeedbackOnWork(ctx context.Context, studentWorkID int64) ([]models.PRReviewCommentResponse, error) {
+func (db *DB) GetFeedbackOnWork(ctx context.Context, studentWorkID int) ([]models.PRReviewCommentResponse, error) {
 	query := `SELECT student_work_id, rubric_item_id, github_username, file_path, file_line, fc.created_at, point_value, explanation
 	FROM feedback_comment fc
 	JOIN rubric_items ri ON fc.rubric_item_id = ri.id
@@ -48,7 +48,7 @@ func (db *DB) GetFeedbackOnWork(ctx context.Context, studentWorkID int64) ([]mod
 }
 
 // create a new feedback comment (ad-hoc: also create a rubric item simultaneously)
-func (db *DB) CreateFeedbackComment(ctx context.Context, TAUserID int64, studentWorkID int64, comment models.PRReviewCommentResponse) error {
+func (db *DB) CreateFeedbackComment(ctx context.Context, TAUserID int64, studentWorkID int, comment models.PRReviewCommentResponse) error {
 	_, err := db.connPool.Exec(ctx,
 		`WITH ri AS
 			(INSERT INTO rubric_items (point_value, explanation) VALUES ($1, $2) RETURNING id)
@@ -66,7 +66,7 @@ func (db *DB) CreateFeedbackComment(ctx context.Context, TAUserID int64, student
 }
 
 // create a new feedback comment (attach existing rubric item)
-func (db *DB) AttachRubricItemToFeedbackComment(ctx context.Context, TAUserID int64, studentWorkID int64, comment models.PRReviewCommentResponse) error {
+func (db *DB) AttachRubricItemToFeedbackComment(ctx context.Context, TAUserID int64, studentWorkID int, comment models.PRReviewCommentResponse) error {
 	if comment.RubricItemID == nil {
 		return errors.New("no rubric item id given")
 	}
